+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2022.08.26`.
+ * Upgraded stable Dart analysis SDK to `2.18.0`.
+ * Upgraded preview Dart analysis SDK to `2.18.0`.
 
 ## `20220825t132100-all`
  * NOTE: `search` stopped populating `highlightedHit` field.